(#353) Merged smargons into one
