+++ src/artemis/devices/fast_grid_scan_composite.py
@@ -1,7 +1,7 @@
 from ophyd import Component, Device, FormattedComponent
 
 from artemis.devices.fast_grid_scan import FastGridScan
-from artemis.devices.motors import I03Smargon
+from artemis.devices.I03Smargon import I03Smargon
 from artemis.devices.slit_gaps import SlitGaps
 from artemis.devices.synchrotron import Synchrotron
 from artemis.devices.undulator import Undulator
