+++ src/artemis/devices/unit_tests/test_gridscan.py
@@ -12,7 +12,7 @@ from artemis.devices.fast_grid_scan import (
     set_fast_grid_scan_params,
     time,
 )
-from artemis.devices.motors import I03Smargon
+from artemis.devices.I03Smargon import I03Smargon
 from artemis.utils import Point3D
